Switch to 8 messages max to keep the size of the program account below 10k. This is a restriction when generating accounts via inner instructions that we will need to work around.
Details:

Initially I was creating the account as an inner instruction, i.e. in the program that owns the program address, but this limited me to a size of 10kb:

"data size limited to 10240 in inner instructions"

So I tried to create the account via a client, using SystemProgram.createAccount. However this gives me a "Signature verification failed" error unless I sign with the new account secret key, which I don't have, because it's a program address. Is what I am doing impossible?